backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
numpy
rasterio
joblib
scikit-learn
xgboost
lightgbm
catboost
requests
